+++ source/runner/templates/_tests/_bitbucketCloudTemplate.test.ts
@@ -4,12 +4,15 @@ import {
   warnResults,
   failsResults,
   messagesResults,
+  customIconMessagesResults,
+  multipleMessagesResults,
   markdownResults,
   summaryResults,
   multipleSummaryResults,
 } from "../../_tests/fixtures/ExampleDangerResults"
 import { dangerSignaturePostfix, template, inlineTemplate } from "../bitbucketCloudTemplate"
 import { DangerResults } from "../../../dsl/DangerResults"
+import * as utils from "../../DangerUtils"
 
 const commitID = "e70f3d6468f61a4bef68c9e6eaba9166b096e23c"
 
@@ -18,11 +21,22 @@ const warningEmoji = "⚠️"
 const messageEmoji = "✨"
 
 describe("generating messages for BitBucket cloud", () => {
+  let complimentMock: jest.Mock<string, []>
+  beforeEach(() => {
+    complimentMock = jest.spyOn(utils, "compliment").mockReturnValue("Well done.")
+  })
+
+  afterEach(() => {
+    complimentMock.mockRestore()
+  })
+
   it("shows no sections for empty results", () => {
     const issues = template("blankID", emptyResults, commitID)
     expect(issues).not.toContain("Fails")
     expect(issues).not.toContain("Warnings")
     expect(issues).not.toContain("Messages")
+    expect(issues).toContain("Well done.")
+    expect(complimentMock).toBeCalledTimes(1)
   })
 
   it("shows no sections for results without messages", () => {
@@ -30,18 +44,39 @@ describe("generating messages for BitBucket cloud", () => {
     expect(issues).not.toContain("Fails")
     expect(issues).not.toContain("Warnings")
     expect(issues).not.toContain("Messages")
+    expect(issues).not.toContain("Well done.")
+    expect(complimentMock).not.toBeCalled()
   })
 
   it("Shows the failing messages in a section", () => {
     const issues = template("blankID", failsResults, commitID)
     expect(issues).toContain("Fails")
     expect(issues).not.toContain("Warnings")
+    expect(issues).not.toContain("Well done.")
+    expect(complimentMock).not.toBeCalled()
   })
 
   it("Shows the warning messages in a section", () => {
     const issues = template("blankID", warnResults, commitID)
     expect(issues).toContain("Warnings")
     expect(issues).not.toContain("Fails")
+    expect(issues).not.toContain("Well done.")
+    expect(complimentMock).not.toBeCalled()
+  })
+
+  it("Shows custom icon for message", () => {
+    const issues = template("blankID", customIconMessagesResults, commitID)
+    expect(issues).toContain("Messages")
+    expect(issues).toContain("📝")
+    expect(issues).not.toContain(messageEmoji)
+    expect(issues).not.toContain("Warnings")
+    expect(issues).not.toContain("Fails")
+    expect(issues).toContain("Well done.")
+    expect(complimentMock).toBeCalled()
+  })
+
+  it("Mixed icon messages match snapshot", () => {
+    expect(template("blankID", multipleMessagesResults, commitID)).toMatchSnapshot()
   })
 
   it("summary result matches snapshot, with a commit", () => {
@@ -63,6 +98,10 @@ describe("generating messages for BitBucket cloud", () => {
   it("shows a postfix message with no commit ID if not provided", () => {
     expect(template("blankID", emptyResults)).toContain(dangerSignaturePostfix({} as DangerResults))
   })
+
+  it("empty result matches snapshot, with a compliment", () => {
+    expect(template("blankID", emptyResults, commitID)).toMatchSnapshot()
+  })
 })
 
 describe("generating inline messages", () => {
@@ -83,6 +122,23 @@ describe("generating inline messages", () => {
   it("Shows the message", () => {
     const issues = inlineTemplate("blankID", messagesResults, "File.swift", 5)
     expect(issues).toContain(`- ${messageEmoji} Message`)
+    expect(issues).not.toContain(`- ${noEntryEmoji}`)
+    expect(issues).not.toContain(`- ${warningEmoji}`)
+  })
+
+  it("Shows message with custom icon", () => {
+    const issues = inlineTemplate("blankID", customIconMessagesResults, "File.swift", 10)
+    expect(issues).toContain("- 📝 Message with custom icon")
+    expect(issues).not.toContain(`- ${messageEmoji}`)
+    expect(issues).not.toContain(`- ${noEntryEmoji}`)
+    expect(issues).not.toContain(`- ${warningEmoji}`)
+  })
+
+  it("Shows mixed messages", () => {
+    const issues = inlineTemplate("blankID", multipleMessagesResults, "File.swift", 10)
+    expect(issues).toContain("- 📝 Message with custom icon")
+    expect(issues).toContain("- 🔔 Message with custom icon2")
+    expect(issues).toContain(`- ${messageEmoji} Test message`)
     expect(issues).not.toContain(`- ${noEntryEmoji}`)
     expect(issues).not.toContain(`- ${warningEmoji}`)
   })