+++ CHANGELOG.md
@@ -16,6 +16,9 @@
 <!-- Your comment below this -->
 
 - Fix Github Actions documentation - [@ravanscafi]
+- Improve Performance by Caching BitBucket Cloud Commits - [@hellocore]
+- Add compliment message to comment template on Bitbucket Cloud - [@hellocore]
+- Add option to set custom icon in messages - [@rohit-smpx]
 
 # 9.0.4
 
@@ -23,7 +26,7 @@
 
 # 9.0.3
 
-- Add support OAuth for BitBucket Cloud - [@HelloCore]
+- Add support OAuth for BitBucket Cloud - [@hellocore]
 - Allow `handleResults` to be called without a `git` object - [@jtreanor]
 
 # 9.0.2
@@ -32,11 +35,11 @@
 
 # 9.0.1
 
-- Fixed incorrect main comment template on Bitbucket Cloud - [@HelloCore]
+- Fixed incorrect main comment template on Bitbucket Cloud - [@hellocore]
 
 # 9.0.0
 
-- Add BitBucket Cloud & BitBucket Pipelines support - [@HelloCore]
+- Add BitBucket Cloud & BitBucket Pipelines support - [@hellocore]
 - Add GitLab missing states - [@f-meloni]
 - Fixes incorrect slug for builds from forks on Codefresh - [@stevenp]
 
@@ -1635,6 +1638,7 @@ Not usable for others, only stubs of classes etc. - [@orta]
 [@fwal]: https://github.com/fwal
 [@happylinks]: https://github.com/happylinks
 [@hongrich]: https://github.com/hongrich
+[@hellocore]: https://github.com/HelloCore
 [@imorente]: https://github.com/imorente
 [@joarwilk]: https://github.com/joarwilk
 [@johansteffner]: https://github.com/johansteffner
@@ -1651,6 +1655,7 @@ Not usable for others, only stubs of classes etc. - [@orta]
 [@peterjgrainger]: https://github.com/peterjgrainger
 [@randak]: https://github.com/randak
 [@ravanscafi]: https://github.com/ravanscafi
+[@rohit-smpx]: https://github.com/rohit-smpx
 [@sajjadzamani]: https://github.com/sajjadzamani
 [@sebinsua]: https://github.com/sebinsua
 [@sgtcoolguy]: https://github.com/sgtcoolguy